requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
confusable-homoglyphs==3.2.0
cycler==0.10.0
decorate-url==1.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==2.1.2
django-appconf==1.0.3
django-awesome-avatar==1.1.3
django-axes==5.0.20
django-betterforms==1.2
django-bootstrap==0.2.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-datepicker-widget==0.1
django-bootstrap-static==4.2.1
django-bootstrap3==12.0.0
django-bootstrap4==1.0.1
django-braces==1.13.0
django-datatable==0.3.1
django-extensions==2.2.5
django-filter==2.2.0
django-glrm==1.1.3
django-import-export==2.3.0
django-ipware==2.1.0
django-jquery==3.1.0
django-multiselectfield==0.1.10
django-pdb==0.6.2
django-range-slider==0.2.4
django-request==1.5.5
django-rest-framework==0.1.0
django-smart-selects==1.5.4
django-tempus-dominus==5.1.2.13
django-toast-messages==0.1
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
et-xmlfile==1.0.1
httpie==1.0.3
idna==2.8
jdcal==1.4.1
kiwisolver==1.1.0
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.1.2
mysqlclient==2.0.1
names==0.3.0
numpy==1.17.4
odfpy==1.4.1
openpyxl==3.0.2
pandas==0.25.3
Pillow==6.2.1
psycopg2==2.8.6
Pygments==2.5.2
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
reportlab==3.5.67
requests==2.22.0
retrying==1.3.3
six==1.11.0
soupsieve==1.9.5
sqlparse==0.4.1
tablib==2.0.0
typing-extensions==3.10.0.0
urllib3==1.25.7
